08.12 - final refactorings (step 4 - refactor the carts feature) - create a carts package. move all the cart related code to the cart package
