+++ src/main/java/com/jparkkennaby/store/carts/CartEmptyException.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.exceptions;
+package com.jparkkennaby.store.carts;
 
 public class CartEmptyException extends RuntimeException {
     public CartEmptyException() {
