+++ src/main/java/com/jparkkennaby/store/carts/CartItem.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.entities;
+package com.jparkkennaby.store.carts;
 
 import java.math.BigDecimal;
