+++ src/main/java/com/jparkkennaby/store/carts/CartItemDto.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.dtos;
+package com.jparkkennaby.store.carts;
 
 import java.math.BigDecimal;
 
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 public class CartItemDto {
-    private CartProductDto product;
+    private ProductDto product;
     private int quantity;
     private BigDecimal totalPrice;
