+++ src/main/java/com/jparkkennaby/store/carts/CartMapper.java
@@ -1,12 +1,7 @@
-package com.jparkkennaby.store.mappers;
+package com.jparkkennaby.store.carts;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-
-import com.jparkkennaby.store.dtos.CartDto;
-import com.jparkkennaby.store.dtos.CartItemDto;
-import com.jparkkennaby.store.entities.Cart;
-import com.jparkkennaby.store.entities.CartItem;
 
 @Mapper(componentModel = "spring") // so spring is made aware and can make beans at runtime
 public interface CartMapper {
