+++ src/main/java/com/jparkkennaby/store/carts/CartNotFoundException.java
@@ -1,4 +1,4 @@
-package com.jparkkennaby.store.exceptions;
+package com.jparkkennaby.store.carts;
 
 public class CartNotFoundException extends RuntimeException {
     public CartNotFoundException() {
