+++ src/main/java/com/jparkkennaby/store/entities/Order.java
@@ -5,6 +5,7 @@ import java.time.LocalDateTime;
 import java.util.LinkedHashSet;
 import java.util.Set;
 
+import com.jparkkennaby.store.carts.Cart;
 import com.jparkkennaby.store.users.User;
 
 import jakarta.persistence.*;
