+++ src/main/java/com/jparkkennaby/store/payments/CheckoutService.java
@@ -4,12 +4,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.jparkkennaby.store.auth.AuthService;
+import com.jparkkennaby.store.carts.CartEmptyException;
+import com.jparkkennaby.store.carts.CartNotFoundException;
+import com.jparkkennaby.store.carts.CartRepository;
+import com.jparkkennaby.store.carts.CartService;
 import com.jparkkennaby.store.entities.Order;
-import com.jparkkennaby.store.exceptions.CartEmptyException;
-import com.jparkkennaby.store.exceptions.CartNotFoundException;
-import com.jparkkennaby.store.repositories.CartRepository;
 import com.jparkkennaby.store.repositories.OrderRepository;
-import com.jparkkennaby.store.services.CartService;
 
 import lombok.RequiredArgsConstructor;
