+++ src/main/java/com/jparkkennaby/store/services/SeedService.java
@@ -11,12 +11,12 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import com.github.javafaker.Faker;
+import com.jparkkennaby.store.carts.CartRepository;
 import com.jparkkennaby.store.exceptions.DatabaseSeedingException;
 import com.jparkkennaby.store.products.Category;
 import com.jparkkennaby.store.products.CategoryRepository;
 import com.jparkkennaby.store.products.Product;
 import com.jparkkennaby.store.products.ProductRepository;
-import com.jparkkennaby.store.repositories.CartRepository;
 import com.jparkkennaby.store.repositories.OrderRepository;
 import com.jparkkennaby.store.users.Role;
 import com.jparkkennaby.store.users.User;
